requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor>=3.3
beanie>=1.25
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
mongomock-motor>=0.0.29
